Remove 2-key bottom row combos for tab and enter
I can reliably do the bottom row 3-key combos for tab and enter
on my Gamma Omega TC36K keyboard, and my QMK/Vial Corne, but less-so
on the Apple laptop with Karabiner-Elements (timing difference?).

Being splayed, the top row 3-key combos for escape and backspace
are harder on the Gamma Omega, so keep the 2-key versions for now.

This solves a clash with the bottom row 2-key psuedo-layer combos
in Naginata Style.